Add wizard creation command (#36)
* Move exception handling code to common

* Set up framework for wizard creation

The command will eventually create the services and collection for
the Enhanced Information Retrieval pattern in a single command.  For
the time being this will be referred to as 'turtle' until we've established
the the final name.

Later commits will add more to the wizard command.

* Add '--wait' option to cluster creation

* Add a timeout for when waiting for a cluster takes too long

If something goes wrong when creating a cluster, a cluster can
forever stay in the "NOT_AVAILABLE" state.  We should have a timeout
so that a user doesn't end up waiting forever.

* Test some of the wizard functions

* Add in the remaining actions for the wizard command

* Rename references to 'turtle' creation command as 'wizard'

* Bump the delay when waiting for cluster creation

* Move 'create wizard' functionality to a new command 'assemble'

* Actually use the enterprise flag when running the 'assemble' command

* Allow users to specify the cluster size when running 'assemble'